AGENTS/smolagents-rag/requirements.txt
smolagents
langchain
datasets
langchain-community
rank_bm25
python-dotenv
